Merge in compatability mode with libmagic
The libmagic distribution uses the same package name `magic` as
python-magic, but has an incompatible API.  This change merges in a
copy of libmagic's bindings, wrapped to give deprecation warnings.

This is intended to a) mitigate the short-term pain to users and
packagers who need to figure out which to use, and b) give us a path
to merging the two sets of bindings.

I'd be happy for libmagic to take over this package if we could find a
path to it.